Crate oak_functions_abi

Type, constant and Wasm host function definitions for the Oak-Functions application binary interface (ABI).

Re-exports

pub use crate::proto::ExtensionHandle;

Modules

proto

Structs

ReportMetricRequest

Holds the label and the value to report a metric.

ReportMetricResponse

Request

See REQUEST_RESPONSE_ENCODING.MD in the crate root.

Response

See REQUEST_RESPONSE_ENCODING.MD in the crate root.

StorageGetItemResponse

Holds the optional value from the storage.

TfModelInferResponse

Enums

ReportMetricError

StatusCode

See REQUEST_RESPONSE_ENCODING.MD in the crate root.

TestingRequest

TestingResponse

TfModelInferError

Functions

invoke⚠

See invoke.

read_request⚠

See read_request.

write_response⚠

See write_response.