Module oak_functions_abi::proto

Structs

ConfigurationInfo

Configuration information that should be provided to the client.

Entry

An individual entry to be made available for lookup to an Oak Function.

Inference

The inference from a TensorFlow model, containing an inference vector of floats, and a shape vector specifying the dimensions of the inference vector.

PrivateMetricsConfig

Configuration for differentially private metrics.

Request

Represents a request sent to an Oak Functions application. It wraps a byte array that the Wasm module can interpret as a request it can handle.

Response

Represents the response from an Oak Functions application. This is included in the body of the HTTP response sent to the client.

ServerPolicy

Enums

OakStatus

Status values exchanged as i32 values across the Node Wasm interface.

StatusCode

Status codes for the response.